train/requirements.txt
scipy==1.1.0
nipype==1.2.0
tqdm==4.32.2
nibabel==2.4.1
Keras==2.2.4
dicom2nifti==2.1.5
h5py==2.9.0
imageio==2.5.0
ipywidgets==7.5.0
opencv_python_headless==4.1.2.30
matplotlib==3.0.3
numpy==1.16.4
scikit_image==0.15.0
ConfigArgParse==0.14.0
pydicom==1.3.0
sigpy==0.1.14
expiringdict==1.1.4
dicom==0.9.9.post1
scikit_learn==0.21.3
tensorflow-gpu==1.12.0
deepbrain
pandas
